+++ Application/EmptySource/Source/include/Math/IntVector3.inl
@@ -173,10 +173,10 @@ FORCEINLINE IntVector3& IntVector3::operator/=(const int& Value) {
 	return *this;
 }
 
-inline IntVector3 operator*(float Value, const IntVector3 & Vector) {
+inline IntVector3 operator*(int Value, const IntVector3 & Vector) {
 	return IntVector3(Value * Vector.x, Value * Vector.y, Value / Vector.z);
 }
 
-inline IntVector3 operator/(float Value, const IntVector3 & Vector) {
+inline IntVector3 operator/(int Value, const IntVector3 & Vector) {
 	return IntVector3(Value / Vector.x, Value / Vector.y, Value / Vector.z);
 }
